requirements.txt
fastapi>=0.115,<1.0
strawberry-graphql[fastapi]>=0.275,<1.0
uvicorn[standard]>=0.35,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
